requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==4.2.3
django-environ==0.10.0
django-filter==23.2
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.6
fonttools==4.53.0
GDAL==3.4.1
gunicorn==20.1.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.3.7
MarkupSafe==2.1.5
matplotlib==3.7.2
numpy==2.0.0
packaging==24.1
pandas==2.0.3
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
